Refactor/modernize

+++ src/main/java/com/delegrego/exemplo_spring_boot_2/dto/departamento/response/DepartamentoResponseDto.java
@@ -1,7 +1,5 @@
 package com.delegrego.exemplo_spring_boot_2.dto.departamento.response;
 
-import java.math.BigInteger;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class DepartamentoResponseDto {
 
-	private BigInteger idDepartamento;
+	private Long idDepartamento;
 
 	private String nmDepartamento;
