+++ src/main/java/com/delegrego/exemplo_spring_boot_2/dto/funcionario/response/FuncionarioResponseDto.java
@@ -1,7 +1,6 @@
 package com.delegrego.exemplo_spring_boot_2.dto.funcionario.response;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 import com.delegrego.exemplo_spring_boot_2.dto.departamento.response.DepartamentoResponseDto;
@@ -20,7 +19,7 @@ import lombok.ToString;
 @ToString
 public class FuncionarioResponseDto {
 
-	private BigInteger idFuncionario;
+	private Long idFuncionario;
 
 	private String nome;
