+++ src/main/java/com/delegrego/exemplo_spring_boot_2/entity/DepartamentoEntity.java
@@ -1,6 +1,5 @@
 package com.delegrego.exemplo_spring_boot_2.entity;
 
-import java.math.BigInteger;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -31,7 +30,7 @@ public class DepartamentoEntity {
 	// IDENTITY é o tipo utilizado pelo MySQL
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_departamento")
-	private BigInteger idDepartamento;
+	private Long idDepartamento;
 
 	// Restringe o atributo para não ser nulo e ter limite de 50 caracteres
 	@Column(name = "nm_departamento", nullable = false, length = 50)
